#131 refactor: Category DTO 분리
