--- src/main/java/com/example/testypie/domain/category/dto/CategoryCreateRequestDTO.java
@@ -1,3 +1,0 @@
-package com.example.testypie.domain.category.dto;
-
-public record CategoryCreateRequestDTO(String name, String parent) {}
