+++ src/main/java/com/example/testypie/domain/category/dto/request/CreateCategoryRequestDTO.java
@@ -1,0 +1,3 @@
+package com.example.testypie.domain.category.dto.request;
+
+public record CreateCategoryRequestDTO(String name, String parent) {}
