+++ src/main/java/com/example/testypie/domain/category/service/CategoryService.java
@@ -1,7 +1,7 @@
 package com.example.testypie.domain.category.service;
 
-import com.example.testypie.domain.category.dto.CategoryCreateRequestDTO;
-import com.example.testypie.domain.category.dto.CategoryCreateResponseDTO;
+import com.example.testypie.domain.category.dto.request.CreateCategoryRequestDTO;
+import com.example.testypie.domain.category.dto.response.CreateCategoryResponseDTO;
 import com.example.testypie.domain.category.entity.Category;
 import com.example.testypie.domain.category.repository.CategoryRepository;
 import com.example.testypie.global.exception.ErrorCode;
@@ -15,12 +15,12 @@ public class CategoryService {
 
   private final CategoryRepository categoryRepository;
 
-  public CategoryCreateResponseDTO createCategory(CategoryCreateRequestDTO req) {
+  public CreateCategoryResponseDTO createCategory(CreateCategoryRequestDTO req) {
     Category parentName = categoryRepository.findByName(req.parent());
     Category category =
         Category.builder().name(req.name()).depth((long) 1).parent(parentName).build();
     Category saveCategory = categoryRepository.save(category);
-    return CategoryCreateResponseDTO.of(saveCategory);
+    return CreateCategoryResponseDTO.of(saveCategory);
   }
 
   public Category getParentCategory(String parentCategory_name) {
